requirements.txt
beautifulsoup4==4.12.2
google==3.0.0
grpcio==1.59.2
imutils==0.5.4
numpy==1.26.2
opencv-python==4.8.1.78
protobuf==4.25.1
soupsieve==2.5
torch==2.1.1
torchvision==0.16.1